+++ channels/audin/client/audin_main.c
@@ -113,7 +113,7 @@ static UINT audin_channel_write_and_free(AUDIN_CHANNEL_CALLBACK* callback, wStre
 		return ERROR_INTERNAL_ERROR;
 
 	Stream_SealLength(out);
-	WINPR_ASSERT(Stream_Length(out) <= ULONG_MAX);
+	WINPR_ASSERT(Stream_Length(out) <= UINT32_MAX);
 	const UINT error = callback->channel->Write(callback->channel, (ULONG)Stream_Length(out),
 	                                            Stream_Buffer(out), NULL);
