+++ winpr/libwinpr/sspi/NTLM/ntlm_av_pairs.c
@@ -135,7 +135,7 @@ ULONG ntlm_av_pair_list_length(NTLM_AV_PAIR* pAvPairList, size_t cbAvPairList)
 		return 0;
 
 	const size_t size = ((PBYTE)pAvPair - (PBYTE)pAvPairList) + sizeof(NTLM_AV_PAIR);
-	WINPR_ASSERT(size <= ULONG_MAX);
+	WINPR_ASSERT(size <= UINT32_MAX);
 	WINPR_ASSERT(size >= 0);
 	return (ULONG)size;
 }
